% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{add_columns_for_lookup}
\alias{add_columns_for_lookup}
\title{Afegim columnes per tal de fer la recerca a Maxmind.}
\usage{
add_columns_for_lookup(pVerbose, pDataFrame)
}
\value{
Obtenim el DataFrame que busquem amb columnes addicionals.
}
\description{
Afegim columnes per tal de fer la recerca a Maxmind.
}
\details{
Afegim les columnes necessàries per tal de fer la recerca de les IPs del dataset al dataset de Maxmind.
}
\examples{
--
}
