% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{addIPgeolocation}
\alias{addIPgeolocation}
\title{Añadir la Geolocalización de las IPs al dataframe}
\usage{
addIPgeolocation(ips = "", df.maxmind = data.frame(), boost = FALSE)
}
\value{
Dataframe final
}
\description{
Añadir la Geolocalización de las IPs al dataframe
}
\details{
Añadimos la geolocalización de las IPS al dataframe definitivo a partir del dataset
         de maxmind
}
\examples{
\dontrun{
geoips <- addIPgeolocation(ips = c("8.8.8.8", "147.81.23.1"),
                           df.maxmind = download.maxmind())
}
}
