% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{create_directory}
\alias{create_directory}
\title{Creates a directory.}
\usage{
create_directory(pVerbose, pDirectory)
}
\arguments{
\item{pVerbose}{A boolean (TRUE or FALSE).}

\item{pDirectory}{Full path of the directory you want to create.}
}
\value{
nothing.
}
\description{
Creates a directory.
}
\details{
Creates a directory.
}
