% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{decompress_multifile}
\alias{decompress_multifile}
\title{Decompresses a compressed file which contains multiple files}
\usage{
decompress_multifile(pVerbose, pFile.name, pCompress.type, pDelete.file)
}
\arguments{
\item{pVerbose}{A boolean (TRUE or FALSE).}

\item{pFile.name}{filename (including the full path) of the file to be decompressed.}

\item{pCompress.type}{It indicates the type of compressed file.}

\item{pDelete.file}{A boolean (TRUE or FALSE) indicating if the compressed file has to be deleted after being decompressed.}
}
\value{
nothing.
}
\description{
Decompresses a compressed file which contains multiple files
}
\details{
Decompresses a compressed file which contains multiple files.
}
\examples{
--
}
