% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{download_maxmind}
\alias{download_maxmind}
\title{Download maxmind. Descarreguem de la web maxmind.com una base de dades amb la geolocalització de les
IPs públiques mundials que té registrades.}
\usage{
download_maxmind(pVerbose, pOverwrite.data, pTesting)
}
\value{
Obtenim un dataframe amb la posició GPS de les IPs resgistrades.
}
\description{
Download maxmind. Descarreguem de la web maxmind.com una base de dades amb la geolocalització de les
IPs públiques mundials que té registrades.
}
\details{
Descarrega un fitxer maxmind de la web de maxmind.
}
\examples{
--
}
