% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{get_subset_rows}
\alias{get_subset_rows}
\title{Generem una part del dataframe, reduït a un cert nombre de files. El nombre de files es genera
de manera al·leatòria a partir de pData.}
\usage{
get_subset_rows(pVerbose, pData, pNumRows)
}
\value{
Obtenim el df que busquem amb un nombre de files indicat per pNumRows provinent de pData.
}
\description{
Generem una part del dataframe, reduït a un cert nombre de files. El nombre de files es genera
de manera al·leatòria a partir de pData.
}
\details{
Selects (randomly) a number of rows from pData.
}
\examples{
--
}
