% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{log_msg}
\alias{log_msg}
\title{Logs a message to the Log. If a filename is passed as parameter, it will also print the message into the specified log file.}
\usage{
log_msg(pVerbose, pMessage, pLogFile)
}
\arguments{
\item{pVerbose}{A boolean (TRUE or FALSE).}

\item{pMessage}{The message you want to show in the console.}
}
\value{
nothing.
}
\description{
Logs a message to the Log. If a filename is passed as parameter, it will also print the message into the specified log file.
}
\details{
It's very useful to log messages through a function like this, as it will save you time when developing. You will have to just change the name of the variable 'pVerbose' to log or stop logging messages.
}
\examples{
--
}
