% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{lookup_to_maxmind}
\alias{lookup_to_maxmind}
\title{Lookup to Maxmind to get geolocation (Lat, Lon, Accuracy) of IP_Ori and IP_Dest.}
\usage{
lookup_to_maxmind(pVerbose, pAttacks, pMaxMind)
}
\value{
DataFrame with additional columns (Lat, Lon, Accuracy) for IP_Ori and IP_Dest
}
\description{
Lookup to Maxmind to get geolocation (Lat, Lon, Accuracy) of IP_Ori and IP_Dest.
}
\details{
Adds the geolocation columns into the DataFrame, by looking up to Maxmind through srcip_num and dstip_num.
}
\examples{
--
}
