% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{main}
\alias{main}
\title{Main function, which calls the rest of functions}
\usage{
main()
}
\value{
Nothing
}
\description{
Main function, which calls the rest of functions
}
\details{
Main function, which calls the rest of functions.
}
\examples{
--
}
