% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{add_columns_for_lookup}
\alias{add_columns_for_lookup}
\title{Add columns for lookup}
\usage{
add_columns_for_lookup(verbose, pData)
}
\value{
DataFrame with additional columns
}
\description{
Add columns for lookup
}
\details{
Adds the required columns that will be used for the lookup with Maxmind.
}
\examples{
--
}
