% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{download_maxmind}
\alias{download_maxmind}
\title{Download maxmind}
\usage{
download_maxmind()
}
\value{
Returns a df with the GPS position of each registered network
}
\description{
Download maxmind
}
\details{
Downloads maxmind file MaxMind site.
}
\examples{
--
}
