% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{get_subset_rows}
\alias{get_subset_rows}
\title{Get subset of rows}
\usage{
get_subset_rows(pData, pNumRows)
}
\value{
Returns a df with only the requested pNumRows from pData
}
\description{
Get subset of rows
}
\details{
Selects (randomly) a number of rows from pData.
}
\examples{
--
}
