% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.lib.R
\name{main}
\alias{main}
\title{Main function, which calls the rest of functions}
\usage{
main(verbose = TRUE, overwrite.data = FALSE, testing = FALSE,
  scope.test = 10)
}
\value{
Retorna un resum del dataframe final
}
\description{
Main function, which calls the rest of functions
}
\details{
Main function, funció principal que crida la resta de funcions en l'rodre següent:
- download_data


I utilitza, addicionalment, les següents funcions auxiliars:
- log_msg. Imprimeix per consola o a fitxer aquells missatges que el programador troba adequats a l'hora d'executar el codi
- create_directory. Crea, si és necessari, un directori a l'espai de treball per tal de guardar-hi diferents arxius
- decompress_file. Descomprimeix un fitxer que conté un arxiu i elimina el fitxer original
- decompress_multifile. Descomprimeix un fitxer que conté més d'un arxiu i elimina el fitxer original
}
\examples{
--
}
